Fix g++-8 warnings in jeeps.

+++ jeeps/gpssend.cc
@@ -118,7 +118,7 @@ DiagS(void* buf, size_t sz)
 
 int32 GPS_Serial_Write_Packet(gpsdevh* fd, GPS_PPacket& packet)
 {
-  size_t ret;
+  int32 ret;
   const char* m1, *m2;
   GPS_Serial_OPacket ser_pkt;
   UC ser_pkt_data[MAX_GPS_PACKET_SIZE * sizeof(UC)];
